requirements/base.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs==18.2.0
cffi==1.12.2
chardet==3.0.4
discord.py==1.0.0a1699+g69f5a70
Django==2.1.7
idna==2.8
idna-ssl==1.1.0
multidict==4.5.2
psycopg2==2.7.7
pycparser==2.19
PyNaCl==1.2.1
pytz==2018.9
six==1.12.0
typing-extensions==3.7.2
Unipath==1.1
websockets==7.0
yarl==1.3.0
